Broke staking form into several components
